replace dependencies in class
